data/flask/requirements.txt
Flask==0.10.1
Flask-Migrate==1.7.0
Flask-SQLAlchemy==2.1
SQLAlchemy==1.0.12
uWSGI==2.0.12
Werkzeug==0.11.4
argparse==1.2.1
psycopg2==2.6.1
konlpy==0.4.4
JPype1==0.6.1